user-session: add support for linger session
add support for linger session, with a 'linger' flag file that the
user can create inside /home/$user/.runit/ ; the user is also supposed
to create the rusnvdir@user instance in /etc/sv/ (requires root privileges)
and enable it the first time (or wait for it to be created and enabled
the at the first graphic login by the package/pam machinery);
an alternative approach (discarded for now) could be to use a
one-shot/initscript that runs at boot to scan each home directory
for the flag file and make sure to create and enable the rusnvdir
user instance.

+++ user-session/exec-userinstance.sh
@@ -87,6 +87,8 @@ fi
 
 #LOGOUT
 if [ "$PAM_TYPE" = "close_session" ]; then #logout event for $user
+  #check if it's a linger session
+  [ -e "/home/$user/.runit/linger" ] && exit 0
   #disable (stop) our runsvdir-user if it's enabled
   if [ -h "/etc/service/runsvdir@$user" ]; then
     unlink "/etc/service/runsvdir@$user"
